set the flask run host=0.0.0.0 to run in container
I'm not entirely sure why, but without adding this host argument
the flask app wasn't accessible outside the container, even when mapping
the container port to a host port.

Saw this technique in this tutorial on dockerizing flask apps:
https://pyshark.com/containerize-a-flask-application-using-docker/
From the tutorial:
> Important: if you want to test run the app locally, you should have app.run() but when we are looking to containerize the application we switch it to app.run(host=’0.0.0.0′, port=80) because the local port 127.0.0.1 can’t be captured inside a Docker container and Flask also uses port 5000 by default, so we switch it to port 80.

I didn't include the port mapping in app.run() since the default port 5000
seemed to work fine in the container.