upgrade Google terraform provider to 4.30
This commit upgrades the terraform provider to 4.3, which is the latest
version that has support for "google_artifact_registry_repository"
resource.

I'm not sure yet if it makes sense to use Artifact Registry or
Container Registry to host the dot image, but regardless upgrading
this here seems to make sense.

Locally upgraded my provider with:
"terraform init -upgrade"