Remove google cloud provider and resources; configure heroku provider
After some testing, we decided/found that Google Cloud Run wasn't a great
fit for the way dot 3.0 flask app works challenges:
1. Script assumes one running instance/unsure about race conditions with multiple containers/instances
2. Script relies on system memory, Run removes containers and we can't assume the memory will exist with the request count
3. The scheduler background process isn't a good fit for the request-based nature of Run.

Because of these challenges, the script would need to be refactored to work with Cloud Run.
Ideas included using a GCP memory service that each of the request-based containers could use,
to deal with the count being in memory issue.  Secondly, having a Scheduled task service
run the "flush" task at whatever interval against a common memorystore service could help
with the issue of not having available cpu after a request ends/container is terminated.

For the short term, for staging, we're switching to heroku to test the pipeline while
the production plan develops.  Options include refactoring and trying cloud run again,
deploying to a GKE cluster or virtual machine instance.