Change docker & heroku startup cmd to use gunicorn
This commit swithces from using "flask run..." as the default startup command to
useing gunicorn, a more production appropriate web server.

For both dev & in heroku, the --log-level is set to debug to for testing.
After making this switch, the heroku app serves /pixel.gif requests,
but *isn't* running the background task to flush the memory count.