Add dot v2 static files to bucket (/analytics/{VERSION}/pixel.js)
This commit adds the archived version 2 dot files to the repo in a
static directory, and uploads them to the storage bucket at:
/analytics/{VERSION}/pixel.js.

This uses a terraform local list to iterate over hard-coded version numbers.
Probably a more elegant way here to list the directories to copy,
but figured since they are hard coded in this case just hard coding them
should work for us.

Note: terraform will check the checksums of the pixel.js files,
so if those files change terraform *will* notice that and make the
changes (tested that out to check).